requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.1.1
argcomplete==3.1.1
asgiref==3.7.2
attrs==23.1.0
autopep8==2.0.2
billiard==3.6.4.0
bitarray==2.9.2
botocore==1.31.28
build==0.10.0
CacheControl==0.12.14
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.4.1
cytoolz==0.12.2
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2
django-cors-headers==4.0.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
dulwich==0.21.5
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.7.0
eth-keys==0.4.0
eth-rlp==1.0.0
eth-typing==3.5.2
eth-utils==2.3.1
Faker==19.1.0
filelock==3.12.2
frozenlist==1.4.1
greenlet==2.0.2
hexbytes==0.3.1
html5lib==1.1
humanize==4.8.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
installer==0.7.0
itypes==1.2.0
jaraco.classes==3.3.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==23.13.1
kombu==5.3.1
lockfile==0.12.2
lru-dict==1.2.0
MarkupSafe==2.1.3
mccabe==0.7.0
more-itertools==10.1.0
msgpack==1.0.5
multidict==6.0.4
packaging==23.1
parsimonious==0.9.0
pexpect==4.8.0
pipx==1.2.0
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.5.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.1
psycopg2==2.9.9
ptyprocess==0.7.0
pyactiveresource==1.0.1
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.19.1
pyflakes==3.0.1
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pyunormalize==15.1.0
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==2.15.1
redis==5.0.0
referencing==0.30.2
regex==2023.12.25
requests==2.28.2
requests-toolbelt==1.0.0
rlp==4.0.0
rpds-py==0.9.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.6.2
shellingham==1.5.3
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.1
toolz==0.12.0
tornado==6.3.3
trove-classifiers==2023.8.7
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
userpath==1.9.0
vine==5.0.0
virtualenv==20.24.3
wcwidth==0.2.6
web3==6.13.0
webencodings==0.5.1
websockets==12.0
whitenoise==6.6.0
yarl==1.9.4
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
